refactor: ScrollableView designsystem으로 추출
