+++ feature/onboading/src/main/java/com/saeongjima/signup/AgreeScreen.kt
@@ -1,21 +1,13 @@
 package com.saeongjima.signup
 
-import android.view.View
-import android.view.ViewGroup.LayoutParams.MATCH_PARENT
-import android.view.ViewGroup.LayoutParams.WRAP_CONTENT
-import android.widget.LinearLayout
-import android.widget.ScrollView
 import androidx.annotation.StringRes
-import androidx.appcompat.content.res.AppCompatResources
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -30,22 +22,23 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import androidx.compose.ui.window.DialogWindowProvider
+import com.saeongjima.designsystem.component.ScrollableView
+import com.saeongjima.designsystem.component.button.MainButton
 import com.saeongjima.designsystem.component.checkbox.CheckboxWithLabel
+import com.saeongjima.designsystem.theme.Black100
 import com.saeongjima.designsystem.theme.Black200
 import com.saeongjima.designsystem.theme.Black500
 import com.saeongjima.designsystem.theme.Black700
 import com.saeongjima.designsystem.theme.Black950
+import com.saeongjima.designsystem.theme.PointColor1
 import com.saeongjima.login.R
 
 @Composable
@@ -96,7 +89,7 @@ fun AgreeScreen(modifier: Modifier = Modifier, enableButton: (Boolean) -> Unit) 
                     },
             )
         }
-        DrawScrollableView(
+        ScrollableView(
             modifier = modifier
                 .padding(top = 12.dp)
                 .weight(1f)
@@ -135,7 +128,8 @@ fun AgreeScreen(modifier: Modifier = Modifier, enableButton: (Boolean) -> Unit) 
                     },
             )
         }
-        DrawScrollableView(
+
+        ScrollableView(
             modifier = modifier
                 .padding(top = 12.dp)
                 .weight(1f)
@@ -152,7 +146,11 @@ fun AgreeScreen(modifier: Modifier = Modifier, enableButton: (Boolean) -> Unit) 
 }
 
 @Composable
-fun FullTextDialog(text: String, modifier: Modifier = Modifier, onDismissRequest: () -> Unit) {
+private fun FullTextDialog(
+    text: String,
+    modifier: Modifier = Modifier,
+    onDismissRequest: () -> Unit
+) {
     Dialog(
         onDismissRequest = { onDismissRequest() },
         properties = DialogProperties(usePlatformDefaultWidth = false)
@@ -176,7 +174,7 @@ fun FullTextDialog(text: String, modifier: Modifier = Modifier, onDismissRequest
                     tint = Black700,
                 )
             }
-            DrawScrollableView(
+            ScrollableView(
                 modifier = modifier
                     .padding(top = 12.dp)
                     .weight(1f)
@@ -190,41 +188,6 @@ fun FullTextDialog(text: String, modifier: Modifier = Modifier, onDismissRequest
             }
         }
     }
-}
-
-@Composable
-fun DrawScrollableView(modifier: Modifier = Modifier, content: @Composable () -> Unit) {
-    val context = LocalContext.current
-    AndroidView(
-        modifier = modifier.padding(16.dp),
-        factory = {
-            val scrollView = ScrollView(it)
-            val layout = LinearLayout.LayoutParams(MATCH_PARENT, MATCH_PARENT)
-            scrollView.layoutParams = layout
-            scrollView.isVerticalFadingEdgeEnabled = true
-            scrollView.isScrollbarFadingEnabled = false
-            scrollView.isVerticalScrollBarEnabled = true
-            scrollView.scrollBarSize = 12
-            scrollView.scrollBarStyle = View.SCROLLBARS_OUTSIDE_INSET
-            scrollView.isSmoothScrollingEnabled = true
-            scrollView.verticalScrollbarThumbDrawable = AppCompatResources.getDrawable(
-                context,
-                com.saeongjima.designsystem.R.drawable.scroll_bar,
-            )
-            scrollView.addView(
-                ComposeView(it).apply {
-                    setContent {
-                        content()
-                    }
-                },
-            )
-            val linearLayout = LinearLayout(it)
-            linearLayout.orientation = LinearLayout.VERTICAL
-            linearLayout.layoutParams = LinearLayout.LayoutParams(MATCH_PARENT, WRAP_CONTENT)
-            linearLayout.addView(scrollView)
-            linearLayout
-        },
-    )
 }
 
 enum class PolicyText(@StringRes val res: Int) {
